+++ device-minion-runtime/src/WorkerMinion/include/kernel_return.h
@@ -4,10 +4,11 @@
 #include <stdint.h>
 
 /* Kernel return types */
-#define KERNEL_RETURN_SUCCESS            0
-#define KERNEL_RETURN_SELF_ABORT         1
-#define KERNEL_RETURN_SYSTEM_ABORT       2
-#define KERNEL_RETURN_EXCEPTION          3
+#define KERNEL_RETURN_SUCCESS      0
+#define KERNEL_RETURN_SELF_ABORT   1
+#define KERNEL_RETURN_SYSTEM_ABORT 2
+#define KERNEL_RETURN_EXCEPTION    3
+#define KERNEL_RETURN_BUS_ERROR    4
 
 /* Function prototypes */
 int64_t return_from_kernel(int64_t return_value, uint64_t return_type);